inference/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
diffusers>=0.24.0
transformers>=4.35.0
accelerate>=0.25.0
Pillow>=10.0.0
numpy>=1.24.0